+++ edgemanage/edge_manage
@@ -11,6 +11,7 @@ zone file for DNS server
 
 from __future__ import absolute_import
 from edgemanage import const, EdgeManage, StateFile, util
+from edgemanage.monitor import Monitor
 
 import argparse
 import json
@@ -124,7 +125,7 @@ def main(dnet, dry_run, config, state_obj,
                      if i.strip() and not i.startswith("#")]
         logging.info("Edge list is %s", str(edge_list))
 
-    monitor = util.Monitor(edge_list)
+    monitor = Monitor(edge_list)
 
     # Load or create our edge state files
     for edge in edge_list:
